+++ lib/servy/handler.ex
@@ -6,6 +6,9 @@ defmodule Servy.Handler do
 
   @pages_path Path.expand("../../pages", __DIR__)
 
+  import Servy.Plugins, only: [rewrite_path: 1, log: 1, track: 1]
+  import Servy.Parser, only: [parse: 1]
+
   @doc "transforms request into a response"
   def handle(request) do
     request
@@ -16,37 +19,6 @@ defmodule Servy.Handler do
     |> track
     |> format_response
   end
-
-  def track(%{status: 404, path: path} = conv) do
-    IO.puts "Warning #{path} is on the loose"
-    conv
-  end
-  def track(conv), do: conv
-
-  def rewrite_path( %{method: method, path: "/wildlife"} = conv) do
-    %{conv | path: "/wildthings"}
-  end
-
-  def rewrite_path(conv), do: conv
-
-  def log(conv), do: IO.inspect conv
-
-
-  def parse(request) do
-    # get the method and path
-    [method, path, _] =
-      request
-      |> String.split("\n")
-      |> List.first
-      |> String.split(" ")
-
-    # la fin
-    %{ method: method,
-    path: path,
-    status: 0,
-    resp_body: "" }
-  end
-
 
   def route(%{method: "GET", path: "/wildthings"} = conv) do
     %{ conv | status: 200, resp_body: "Bears, Lions, Tigers"}
